Refactor folder/file structure
