+++ component/common/ErrorDialog.js
