+++ component/common/constants.js
