+++ component/common/materialComponents/Thumbnail.js
