+++ component/main/HNNParametersContainer.js
@@ -1,8 +1,8 @@
 import React from 'react'
-import CellParams from "../definition/cellParams/CellParams";
-import NetworkParams from "../definition/networkParams/NetworkParams";
-import Run from "../definition/run/Run";
-import Inputs from "../definition/inputs/Inputs";
+import CellParams from "../parameters/cellParams/CellParams";
+import NetworkParams from "../parameters/networkParams/NetworkParams";
+import Run from "../parameters/run/Run";
+import Inputs from "../parameters/inputs/Inputs";
 
 export default ({ visibility }) => (
   <div style={{ width: "100%", visibility, marginTop: "64px" }}>
