+++ component/parameters/inputs/DynamicThumbnail.js
