+++ component/parameters/networkParams/NetworkParams.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 
 import Metadata from '../../../Metadata';
-import Card from '../../general/materialComponents/Card';
-import Thumbnail from '../../general/materialComponents/Thumbnail';
-import Navigation from '../../general/materialComponents/Navigation';
+import Card from '../../common/materialComponents/Card';
+import Thumbnail from '../../common/materialComponents/Thumbnail';
+import Navigation from '../../common/materialComponents/Navigation';
 
 export default class NetworkParams extends Component {
   static contextType = Metadata;
